refactor: move icon components to icons folder, and update import paths
