+++ src/assets/icons/store.tsx
